feat: add pagination to GET /api/students endpoint and also adding pagination to the student index mvc

+++ Repository/StudentRepository.cs
@@ -27,6 +27,10 @@ namespace demoEFapp.Repositoy
         public void DeleteStudent(int id)
         {
             Student student = _dbContext.Students.FirstOrDefault(x => x.StudentId == id);
+
+            if (student == null)
+                return;
+
             _dbContext.Students.Remove(student);
             _dbContext.SaveChanges();
         }
